feat: support native interceptor
