+++ packages/mocker/src/browser/native.ts
@@ -1,0 +1,17 @@
+import type { MockedModule } from '../registry'
+import type { ModuleMockerInterceptor } from './interceptor'
+import { rpc } from './utils'
+
+export class ModuleMockerNativeInterceptor implements ModuleMockerInterceptor {
+  async register(module: MockedModule): Promise<void> {
+    await rpc('vitest:interceptor:register', module.toJSON())
+  }
+
+  async delete(id: string): Promise<void> {
+    await rpc('vitest:interceptor:delete', id)
+  }
+
+  invalidate(): void {
+    rpc('vitest:interceptor:invalidate')
+  }
+}
