+++ packages/mocker/src/node/index.ts
@@ -3,6 +3,7 @@ export { findMockRedirect } from './redirect'
 export { hoistMocksPlugin, hoistMocks } from './hoistMocksPlugin'
 export { ServerMockResolver } from './resolver'
 export { dynamicImportPlugin } from './dynamicImportPlugin'
+export { interceptorPlugin } from './interceptorPlugin'
 export { mockerPlugin } from './mockerPlugin'
 
 export type {
@@ -12,3 +13,4 @@ export type {
 } from './resolver'
 export type { AutomockPluginOptions } from './automockPlugin'
 export type { HoistMocksPluginOptions, HoistMocksResult } from './hoistMocksPlugin'
+export type { InterceptorPluginOptions } from './interceptorPlugin'
